requirements.txt
tensorflow-cpu==2.10.1
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.13
typing-extensions==4.4.0
numpy<2
